Move PostListItemUploadStatus to it's own file
