+++ WordPress/src/main/java/org/wordpress/android/viewmodel/posts/PostListItemUploadStatus.kt
@@ -1,0 +1,14 @@
+package org.wordpress.android.viewmodel.posts
+
+import org.wordpress.android.fluxc.store.UploadStore.UploadError
+
+data class PostListItemUploadStatus(
+    val uploadError: UploadError?,
+    val mediaUploadProgress: Int,
+    val isUploading: Boolean,
+    val isUploadingOrQueued: Boolean,
+    val isQueued: Boolean,
+    val isUploadFailed: Boolean,
+    val hasInProgressMediaUpload: Boolean,
+    val hasPendingMediaUpload: Boolean
+)
